[IMP] combine module
