+++ budget_source_fund/__manifest__.py
@@ -12,6 +12,7 @@
     "data": [
         "security/ir.model.access.csv",
         "views/budget_source_fund_view.xml",
+        "views/budget_source_fund_group_view.xml",
         "views/budget_control_view.xml",
         "report/source_fund_monitor_report_view.xml",
     ],
