+++ budget_source_fund/models/__init__.py
@@ -1,4 +1,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import budget_source_fund
+from . import budget_source_fund_group
 from . import budget_control
