+++ budget_source_fund/report/source_fund_monitor_report.py
@@ -19,6 +19,7 @@ class SourceFundMonitorReport(models.Model):
         comodel_name="account.analytic.account",
     )
     fund_id = fields.Many2one(comodel_name="budget.source.fund")
+    fund_group_id = fields.Many2one(comodel_name="budget.source.fund.group")
     amount = fields.Float()
     amount_type = fields.Selection(
         selection=lambda self: [("1_budget", "Budget")]
@@ -63,6 +64,7 @@ class SourceFundMonitorReport(models.Model):
                 %s000000000 + a.id as id,
                 '%s,' || a.%s as res_id,
                 a.fund_id as fund_id,
+                sf_group.id as fund_group_id,
                 a.analytic_account_id,
                 '%s' as amount_type,
                 a.credit-a.debit as amount
@@ -86,6 +88,8 @@ class SourceFundMonitorReport(models.Model):
                 from {} a
                 left outer join {} b on a.{} = b.id
                 join budget_source_fund sf on sf.id = a.fund_id
+                join budget_source_fund_group sf_group
+                    on sf_group.id = sf.fund_group_id
             """.format(
                 budget_table,
                 doc_table,
@@ -99,6 +103,7 @@ class SourceFundMonitorReport(models.Model):
             1000000000 + sf.id as id,
             'budget.source.fund,' || sf.id as res_id,
             sf.id as fund_id,
+            sf_group.id as fund_group_id,
             null::integer as analytic_account_id,
             '1_budget' as amount_type,
             null::integer as amount
@@ -108,6 +113,8 @@ class SourceFundMonitorReport(models.Model):
     def _from_budget(self):
         return """
             from budget_source_fund sf
+            join budget_source_fund_group sf_group
+            on sf_group.id = sf.fund_group_id
         """
 
     def _where_budget(self):
