+++ budget_source_fund/report/source_fund_monitor_report_view.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record id="view_budget_monitor_report_tree" model="ir.ui.view">
+    <record id="view_source_fund_monitor_report_tree" model="ir.ui.view">
         <field name="name">source.fund.monitor.report.tree</field>
         <field name="model">source.fund.monitor.report</field>
         <field name="arch" type="xml">
@@ -12,7 +12,7 @@
             </tree>
         </field>
     </record>
-    <record id="view_budget_monitor_report_pivot" model="ir.ui.view">
+    <record id="view_source_fund_monitor_report_pivot" model="ir.ui.view">
         <field name="name">source.fund.monitor.report.pivot</field>
         <field name="model">source.fund.monitor.report</field>
         <field name="arch" type="xml">
@@ -31,6 +31,11 @@
                 <field name="fund_id" string="Source of Fund" />
                 <separator />
                 <group expand="1" string="Group By">
+                    <filter
+                        string="Fund Group"
+                        name="fund_group_id"
+                        context="{'group_by':'fund_group_id'}"
+                    />
                     <filter
                         string="Source of Fund"
                         name="fund_id"
