+++ budget_source_fund/views/budget_source_fund_view.xml
@@ -6,6 +6,7 @@
         <field name="arch" type="xml">
             <tree string="Source of Funds" sample="1">
                 <field name="name" />
+                <field name="fund_group_id" />
             </tree>
         </field>
     </record>
@@ -17,11 +18,22 @@
                 <field name="name" string="Source of Fund" />
                 <separator />
                 <filter
+                    string="Fund Group"
+                    name="fund_group_id"
+                    domain="[('fund_group_id', '!=', False)]"
+                />
+                <filter
                     string="Archived"
                     name="inactive"
                     domain="[('active', '=', False)]"
                 />
-                <group expand="0" string="Group By" />
+                <group expand="0" string="Group By">
+                    <filter
+                        string="Fund Group"
+                        name="fund_group_id"
+                        context="{'group_by':'fund_group_id'}"
+                    />
+                </group>
             </search>
         </field>
     </record>
@@ -43,6 +55,13 @@
                         <h1>
                             <field name="name" />
                         </h1>
+                        <h4>
+                            <field
+                                name="fund_group_id"
+                                placeholder="Group of Fund"
+                                options="{'no_create': True}"
+                            />
+                        </h4>
                     </div>
                     <group>
                         <field name="active" invisible="1" />
