--- budget_source_fund_group/__init__.py
@@ -1,4 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import models
-from . import report
