--- budget_source_fund_group/models/__init__.py
@@ -1,4 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import budget_source_fund_group
-from . import budget_source_fund
