--- budget_source_fund_group/models/budget_source_fund.py
@@ -1,14 +1,0 @@
-# Copyright 2021 Ecosoft Co., Ltd. (http://ecosoft.co.th)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-from odoo import fields, models
-
-
-class BudgetSourceFund(models.Model):
-    _inherit = "budget.source.fund"
-
-    fund_group_id = fields.Many2one(
-        comodel_name="budget.source.fund.group",
-        string="Fund Group",
-        required=True,
-        tracking=True,
-    )
