--- budget_source_fund_group/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Saran Lim. <saranl@ecosoft.co.th>
