--- budget_source_fund_group/readme/DESCRIPTION.rst
@@ -1,4 +1,0 @@
-This module allow create Master Data source of fund.
-there is relation between source of fund and budget control sheet
-for allocated source of fund from one budget control sheet to many source of fund.
-Users can view source of fund monitoring report
