--- budget_source_fund_group/readme/USAGE.rst
@@ -1,7 +1,0 @@
-At first you have to create Source of Funds and select Funds at budget control sheet, it will generate source of fund related with budget control sheet.
-
-* Go to menu Budgeting > Configurations > Source of Funds
-* Create your source of fund and Group of fund (optional)
-* Go to menu Budgeting > Budgeting > Budget Control Sheet
-* Select 1 budget control > select fund that depend on this budget control > Save
-* It will generated source of fund line
