--- budget_source_fund_group/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-access_budget_source_fund_group,access_budget_source_fund_group,model_budget_source_fund_group,budget_control.group_budget_control_manager,1,1,1,1
