--- budget_source_fund_group/views/budget_source_fund_view.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="budget_source_fund_tree" model="ir.ui.view">
-        <field name="name">budget.source.fund.tree</field>
-        <field name="model">budget.source.fund</field>
-        <field name="inherit_id" ref="budget_source_fund.budget_source_fund_tree" />
-        <field name="arch" type="xml">
-            <field name="name" position="after">
-                <field name="fund_group_id" />
-            </field>
-        </field>
-    </record>
-    <record id="budget_source_fund_filter" model="ir.ui.view">
-        <field name="name">budget.source.fund.filter</field>
-        <field name="model">budget.source.fund</field>
-        <field name="inherit_id" ref="budget_source_fund.budget_source_fund_filter" />
-        <field name="arch" type="xml">
-            <xpath expr="//search/separator[1]" position="after">
-                <filter
-                    string="Fund Group"
-                    name="fund_group_id"
-                    domain="[('fund_group_id', '!=', False)]"
-                />
-            </xpath>
-            <xpath expr="//search/group" position="inside">
-                <filter
-                    string="Fund Group"
-                    name="fund_group_id"
-                    context="{'group_by':'fund_group_id'}"
-                />
-            </xpath>
-        </field>
-    </record>
-    <!-- Main source of funds configuration -->
-    <record id="budget_source_fund_form" model="ir.ui.view">
-        <field name="name">budget.source.fund.form</field>
-        <field name="model">budget.source.fund</field>
-        <field name="inherit_id" ref="budget_source_fund.budget_source_fund_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//form//div[hasclass('oe_title')]" position="inside">
-                <h4>
-                    <field
-                        name="fund_group_id"
-                        placeholder="Group of Fund"
-                        options="{'no_create': True}"
-                    />
-                </h4>
-            </xpath>
-        </field>
-    </record>
-</odoo>
