+++ res_program/data/program_data.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo noupdate="1">
+    <record id="program_unconstrained" model="res.program">
+        <field name="name">Unconstrained</field>
+    </record>
+</odoo>
