+++ res_project/models/__init__.py
@@ -1,3 +1,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import res_project
+from . import hr_department
+from . import hr_employee
