+++ res_project/views/res_project_views.xml
@@ -8,9 +8,10 @@
                 <field name="name" />
                 <field name="code" optional="show" />
                 <field name="parent_project" />
-                <field name="user_id" />
+                <field name="project_manager_id" />
                 <field name="date_from" />
                 <field name="date_to" />
+                <field name="department_id" />
                 <field
                     name="state"
                     widget="badge"
@@ -79,7 +80,8 @@
                         <field name="active" invisible="1" />
                         <group name="main_left_project">
                             <field name="parent_project" />
-                            <field name="user_id" />
+                            <field name="project_manager_id" />
+                            <field name="department_id" />
                         </group>
                         <group name="main_right_project">
                             <field name="date_from" />
@@ -90,7 +92,11 @@
                             />
                         </group>
                     </group>
-                    <notebook />
+                    <notebook>
+                        <page name="members" string="Members">
+                            <field name="member_ids" nolabel="1" />
+                        </page>
+                    </notebook>
                     <group name="description" string="Description">
                         <field name="description" nolabel="1" />
                     </group>
@@ -109,13 +115,13 @@
             <search string="Search Project">
                 <field name="name" string="Project" />
                 <field name="parent_project" string="Parent Project" />
-                <field name="user_id" string="Project Manager" />
+                <field name="project_manager_id" string="Project Manager" />
                 <separator name="filter_information" />
-                <filter
+                <!-- <filter
                     string="My Projects"
                     name="own_projects"
-                    domain="[('user_id', '=', uid)]"
-                />
+                    domain="[('project_manager_id', '=', uid)]"
+                /> -->
                 <separator name="filter_state" />
                 <filter
                     string="Draft"
@@ -149,8 +155,8 @@
                 <group expand="0" string="Group By">
                     <filter
                         string="Project Manager"
-                        name="Manager"
-                        context="{'group_by': 'user_id'}"
+                        name="project_manager"
+                        context="{'group_by': 'project_manager_id'}"
                     />
                     <filter
                         string="Parent Project"
