--- res_project_department/__init__.py
@@ -1,3 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import models
