--- res_project_department/models/__init__.py
@@ -1,4 +1,0 @@
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import res_project
-from . import hr_department
