--- res_project_department/readme/DESCRIPTION.rst
@@ -1,26 +1,0 @@
-This module allows to group Project on Department or
-select Department on Project.
-
-Department View
-==============================
-Department A
-
-- Project AA
-- Project BB
-- Project CC
-
-Department B
-
-- Project DD
-- Project EE
-
-
-Project View
-==============================
-Project AA
-
-- Department A
-
-Project BB
-
-- Department A
