--- res_project_department/views/res_project_views.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="view_res_project_tree" model="ir.ui.view">
-        <field name="name">res.project.tree</field>
-        <field name="model">res.project</field>
-        <field name="inherit_id" ref="res_project.view_res_project_tree" />
-        <field name="arch" type="xml">
-            <field name="date_to" position="after">
-                <field name="department_id" />
-            </field>
-        </field>
-    </record>
-    <record id="view_res_project_form" model="ir.ui.view">
-        <field name="name">res.project.form</field>
-        <field name="model">res.project</field>
-        <field name="inherit_id" ref="res_project.view_res_project_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//group[@name='main_left_project']" position="inside">
-                <field name="department_id" />
-            </xpath>
-        </field>
-    </record>
-</odoo>
