--- res_project_member/__manifest__.py
@@ -1,16 +1,0 @@
-# Copyright 2021 Ecosoft Co., Ltd. (http://ecosoft.co.th)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
-{
-    "name": "Project Member",
-    "summary": "Add members on Projects",
-    "version": "14.0.1.0.0",
-    "license": "AGPL-3",
-    "category": "Project",
-    "website": "https://github.com/OCA/account-budgeting",
-    "author": "Ecosoft, Odoo Community Association (OCA)",
-    "depends": ["res_project"],
-    "data": ["views/res_project_views.xml", "views/hr_employee_views.xml"],
-    "maintainers": ["Saran440"],
-    "development_status": "Alpha",
-}
