--- res_project_member/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Saran Lim. <saranl@ecosoft.co.th>
