--- res_project_member/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-This module add member on project related with employee
