--- res_project_member/views/res_project_views.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="view_res_project_form" model="ir.ui.view">
-        <field name="name">res.project.form</field>
-        <field name="model">res.project</field>
-        <field name="inherit_id" ref="res_project.view_res_project_form" />
-        <field name="arch" type="xml">
-            <xpath expr="/form/sheet/notebook" position="inside">
-                <page name="members" string="Members">
-                    <field name="member_ids" nolabel="1" />
-                </page>
-            </xpath>
-        </field>
-    </record>
-</odoo>
